Moved collections, parsers, and data from GSE code to telemetry_tools.
