+++ FoGSE/demos/readRawToRefined_existing_rtd.py
@@ -6,8 +6,8 @@ the RTDs
 from PyQt6 import QtCore
 
 from FoGSE.readBackwards import BackwardsReader
-from FoGSE.parsers.RTDparser import rtdparser
-from FoGSE.collections.RTDCollection import RTDCollection
+from FoGSE.telemetry_tools.parsers.RTDparser import rtdparser
+from FoGSE.telemetry_tools.collections.RTDCollection import RTDCollection
 from FoGSE.demos.readRawToRefined_single_det import Reader
 
 
@@ -103,7 +103,7 @@ class RTDFileReader(Reader):
 
         Returns
         -------
-        `FoGSE.detector_collections.CdTeCollection.CdTeCollection` :
+        `FoGSE.telemetry_tools.collections.CdTeCollection.CdTeCollection` :
             The CdTe collection.
         """
         # take human readable and convert and set to
