+++ FoGSE/demos/readRawToRefined_fake_rtd.py
@@ -7,7 +7,7 @@ from PyQt6 import QtCore
 
 from FoGSE.readBackwards import BackwardsReader
 from FoGSE.fake_foxsi.temp_parser import temp_parser
-from FoGSE.collections.RTDCollection import RTDCollection
+from FoGSE.telemetry_tools.collections.RTDCollection import RTDCollection
 from FoGSE.demos.readRawToRefined_single_det import Reader
 
 
@@ -96,7 +96,7 @@ class RTDFileReader(Reader):
 
         Returns
         -------
-        `FoGSE.detector_collections.CdTeCollection.CdTeCollection` :
+        `FoGSE.telemetry_tools.collections.CdTeCollection.CdTeCollection` :
             The CdTe collection.
         """
         # take human readable and convert and set to
